Conflicts:
	getchanges.bat

+++ install.py
@@ -3,23 +3,28 @@ import zipfile, urllib2
 import platform
 import shutil, tempfile, json
 import errno
+import platform
+from minecriftversion import mc_version, of_file_name, of_json_name, minecrift_version_num, minecrift_build, of_file_extension, of_file_md5, mcp_version, mc_file_md5
 from hashlib import md5  # pylint: disable-msg=E0611
 from optparse import OptionParser
-
 from applychanges import applychanges, apply_patch
 
 
 base_dir = os.path.dirname(os.path.abspath(__file__))
 
-mc_version = "1.7.10"
-of_version = mc_version+"_HD_U_A4"
-of_file_extension = ".jar"
-mcp_version = "mcp908"
+preferredarch = ''
+nomerge = False
 
 try:
     WindowsError
 except NameError:
     WindowsError = OSError
+
+def osArch():
+    if platform.machine().endswith('64'):
+        return '64'
+    else:
+        return '32'
 
 def mkdir_p(path):
     try:
@@ -38,24 +43,41 @@ def get_md5(file):
 
 def download_file(url, target, md5=None):
     name = os.path.basename(target)
-
-    if not os.path.isfile(target):
+    download = True
+    if not is_non_zero_file(target):
+        if os.path.isfile(target):
+            os.remove(target)
+        download = True
+    elif not md5 == None and not md5 == "":
+        if not get_md5(target) == md5:
+            print 'File Exists but bad MD5!: %s [MD5:%s]' % ( os.path.basename(target), get_md5(target) )
+            os.remove(target)
+            download = True
+        else:
+            print 'File Exists: %s [MD5:%s]' % ( os.path.basename(target), get_md5(target) )
+            download = False
+    else:
+        print 'File Exists: %s' % os.path.basename(target)
+        download = False 
+        
+    if download is True:
+        print 'Downloading: %s' % os.path.basename(target)
         try:
             with open(target,"wb") as tf:
                 res = urllib2.urlopen(urllib2.Request( url, headers = {"User-Agent":"Mozilla/5.0"}))
                 tf.write( res.read() )
-            if not md5 == None:
+            if not md5 == None and not md5 == "":
                 if not get_md5(target) == md5:
                     print 'Download of %s failed md5 check, deleting' % name
                     os.remove(target)
                     return False
-            print 'Downloaded %s' % name
         except Exception as e:
             print e
             print 'Download of %s failed, download it manually from \'%s\' to \'%s\'' % (target, url, target)
+            if os.path.isfile(target):
+                os.remove(target)
             return False
-    else:
-        print 'File Exists: %s' % os.path.basename(target)
+
     return True
 
 def download_native(url, folder, name):
@@ -68,21 +90,41 @@ def download_native(url, folder, name):
 
     return True
 
-def download_deps( mcp_dir ):
+def is_non_zero_file(fpath):  
+    return True if os.path.isfile(fpath) and os.path.getsize(fpath) > 0 else False
+    
+def download_deps( mcp_dir, download_mc ):
+
+    mcp_exists = True
     if not os.path.exists(mcp_dir+"/runtime/commands.py "):
-        download_file( "http://mcp.ocean-labs.de/files/archive/"+mcp_version+".zip", mcp_version+".zip" )
+        mcp_exists = False
         try:
-            os.mkdir( mcp_dir )
-            mcp_zip = zipfile.ZipFile( mcp_version+".zip" )
-            mcp_zip.extractall( mcp_dir )
-            import stat
-            astyle = os.path.join(mcp_dir,"runtime","bin","astyle-osx")
-            st = os.stat( astyle )
-            os.chmod(astyle, st.st_mode | stat.S_IEXEC)
+            mcp_zip_file = os.path.join( base_dir,mcp_version+".zip" )
+            if os.path.exists( mcp_zip_file ):
+                os.mkdir( mcp_dir )
+                mcp_zip = zipfile.ZipFile( mcp_zip_file )
+                mcp_zip.extractall( mcp_dir )
+                import stat
+                astyle = os.path.join(mcp_dir,"runtime","bin","astyle-osx")
+                st = os.stat( astyle )
+                os.chmod(astyle, st.st_mode | stat.S_IEXEC)
+                mcp_exists = True
         except:
             pass
-        print("Patching mcp.cfg. ignore \"FAILED\" hunks")
-        apply_patch( mcp_dir, "mcp.cfg.patch", os.path.join(mcp_dir,"conf"))
+            
+    if mcp_exists == False:
+        print "No %s directory or zip file found. Please copy the %s.zip file into %s and re-run the command." % (mcp_dir, mcp_dir, base_dir)
+        exit(1)
+            
+    print("Patching mcp.cfg. Ignore \"FAILED\" hunks")
+    apply_patch( mcp_dir, os.path.join("mcppatches", "mcp.cfg.patch"), os.path.join(mcp_dir,"conf"))
+    
+    client_md5 = os.path.join("mcppatches","client.md5")
+    target_client_md5 = os.path.join(mcp_dir,"temp","client.md5")
+    if not os.path.exists(target_client_md5):
+        mkdir_p( os.path.join(mcp_dir,"temp") )
+        print 'Updating client.md5: copying %s to %s' % (client_md5, target_client_md5)
+        shutil.copy(client_md5,target_client_md5)
 
     jars = os.path.join(mcp_dir,"jars")
 
@@ -98,31 +140,76 @@ def download_deps( mcp_dir ):
     else:
         native = "windows"
 
-
+    flat_lib_dir = os.path.join(base_dir,"lib",mc_version)
+    flat_native_dir = os.path.join(base_dir,"lib",mc_version,"natives",native)
+    mkdir_p( flat_lib_dir )
+    mkdir_p( flat_native_dir )
+     
+    # Get minecrift json file
     json_file = os.path.join(versions,mc_version+".json")
-    shutil.copy( os.path.join("installer",mc_version+".json"),json_file)
+    source_json_file = os.path.join("installer",mc_version+".json")
+    print 'Updating json: copying %s to %s' % (source_json_file, json_file)
+    shutil.copy(source_json_file,json_file)
+    
+    # Use optifine json name for destination dir and jar names
+    optifine_dest_dir = os.path.join(jars,"libraries","optifine","OptiFine",of_json_name )
+    mkdir_p( optifine_dest_dir )
 
-    optifine_dir = os.path.join(jars,"libraries","optifine","OptiFine",of_version )
-    mkdir_p( optifine_dir )
-
-    optifine_url = "http://optifine.net/download.php?f=OptiFine_"+of_version+of_file_extension
-    print 'Downloading Optifine from: %s' % optifine_url
-    download_file( optifine_url, os.path.join( optifine_dir, "OptiFine-"+of_version+".jar" ))
+    print 'Checking Optifine...'
+    optifine_jar = "OptiFine-"+of_json_name+".jar"
+    optifine_dest_file = os.path.join( optifine_dest_dir, optifine_jar )
+ 
+    download_optifine = False
+    optifine_md5 = ''
+    if not is_non_zero_file( optifine_dest_file ):
+        download_optifine = True
+    else:
+        optifine_md5 = get_md5( optifine_dest_file )
+        print 'Optifine md5: %s' % optifine_md5
+        if optifine_md5 != of_file_md5:
+            download_optifine = True
+            print 'Bad MD5!'
+        else:
+            print 'MD5 good!'
+    
+    if download_optifine: 
+        # Use optifine filename for URL
+        optifine_url = "http://optifine.net/download.php?f=OptiFine_"+of_file_name+of_file_extension
+        print 'Downloading Optifine...'
+        if not download_file( optifine_url, optifine_dest_file, of_file_md5 ):
+            print 'FAILED to download Optifine!'
+            sys.exit(1)
+        else:
+            shutil.copy(optifine_dest_file,os.path.join(flat_lib_dir, os.path.basename(optifine_dest_file)))
+            
+    if of_file_md5 == "":
+        optifine_md5 = get_md5( optifine_dest_file )
+        print 'Optifine md5: %s' % optifine_md5
+        sys.exit(0)
 
     json_obj = []
     with open(json_file,"rb") as f:
+        #data=f.read()
+        #print 'JSON File:\n%s' % data
         json_obj = json.load( f )
     try:
         newlibs = []
         for lib in json_obj['libraries']:
+            libname = lib["name"]
             skip = False
             if "rules" in  lib:
                 for rule in lib["rules"]:
                     if "action" in rule and rule["action"] == "allow" and "os" in rule:
                         skip = True
+                        for entry in rule["os"]:
+                            if "name" in entry:
+                                if rule["os"]["name"] == native:
+                                    skip = False
 
             if skip:
+                print 'File: %s\nSkipping due to rules' % libname
                 continue
+                
             group,artifact,version = lib["name"].split(":")
             if "url" in lib:
                 repo = lib["url"]
@@ -141,49 +228,80 @@ def download_deps( mcp_dir ):
                 file32 = os.path.join(jars,"libraries",url32.replace("/",os.sep))
                 mkdir_p(os.path.dirname(file32))
                 download_file( repo + url32, file32 )
+                shutil.copy(file32,os.path.join(flat_lib_dir, os.path.basename(file32)))
                 
                 url64 = url.replace('${arch}', '64')
                 file64 = os.path.join(jars,"libraries",url64.replace("/",os.sep))
                 mkdir_p(os.path.dirname(file64))
-                download_file( repo + url64, file64 )                
+                download_file(repo + url64, file64)
+                shutil.copy(file64,os.path.join(flat_lib_dir, os.path.basename(file64)))                
+
+                # Use preferred architecture to choose which natives to extract.
+                if preferredarch is '32':
+                    print '    Using preferred arch 32bit'
+                    extractnatives( lib, jars, file32, flat_native_dir )
+                else:
+                    print '    Using preferred arch 64bit'
+                    extractnatives( lib, jars, file64, flat_native_dir )                
+               
             else:
                 file = os.path.join(jars,"libraries",url.replace("/",os.sep))
                 mkdir_p(os.path.dirname(file))
-                download_file( repo + url, file )
-
-            if "natives" in lib:
-                folder = os.path.join(jars,"versions",mc_version,mc_version+"-natives")
-                mkdir_p(folder)
-                zip = zipfile.ZipFile(file)
-                for name in zip.namelist():
-                    if not name.startswith('META-INF') and not name.endswith('/'):
-                        out_file = os.path.join(folder, name)
-                        if not os.path.isfile(out_file):
-                            print '    Extracting %s' % name
-                            out = open(out_file, 'wb')
-                            out.write(zip.read(name))
-                            out.flush()
-                            out.close()
-
+                if download_file( repo + url, file ) == True:
+                    shutil.copy(file,os.path.join(flat_lib_dir, os.path.basename(file)))  
+                    extractnatives( lib, jars, file, flat_native_dir )
+                
             newlibs.append( lib )
         json_obj['libraries'] = newlibs
         with open(json_file,"wb+") as f:
             json.dump( json_obj,f, indent=1 )
-    except:
-        pass
+    except Exception as e:
+        print 'ERROR: %s' % e
+        raise
 
-    repo = "https://s3.amazonaws.com/Minecraft.Download/"
-    jar_file = os.path.join(versions,mc_version+".jar")
-    jar_url = repo + "versions/"+mc_version+"/"+mc_version+".jar"
-    download_file( jar_url, jar_file )
+    if download_mc == True:
+        repo = "https://s3.amazonaws.com/Minecraft.Download/"
+        jar_file = os.path.join(versions,mc_version+".jar")
+        jar_url = repo + "versions/"+mc_version+"/"+mc_version+".jar"
+        download_file( jar_url, jar_file, mc_file_md5 )
+        shutil.copy(jar_file,os.path.join(flat_lib_dir, os.path.basename(jar_file))) 
+        
+        if mc_file_md5 == "":
+            mc_md5 = get_md5( jar_file )
+            print '%s md5: %s' % ( os.path.basename(jar_file), mc_md5 )
+            sys.exit(0)	
 
+def extractnatives( lib, jars, file, copydestdir ):
+    if "natives" in lib:
+        folder = os.path.join(jars,"versions",mc_version,mc_version+"-natives")
+        mkdir_p(folder)
+        zip = zipfile.ZipFile(file)
+        #print 'Native extraction: folder: %s, file to unzip: %s' % (folder, file)
+        for name in zip.namelist():
+            if not name.startswith('META-INF') and not name.endswith('/'):
+                out_file = os.path.join(folder, name)
+                print '    Extracting native library %s' % name
+                out = open(out_file, 'wb')
+                out.write(zip.read(name))
+                out.flush()
+                out.close()
+                shutil.copy(out_file,os.path.join(copydestdir, os.path.basename(out_file))) 
 
 def zipmerge( target_file, source_file ):
     out_file, out_filename = tempfile.mkstemp()
     out = zipfile.ZipFile(out_filename,'a')
-    target = zipfile.ZipFile( target_file, 'r')
-    source = zipfile.ZipFile( source_file, 'r' )
+    try:
+        target = zipfile.ZipFile( target_file, 'r')
+    except Exception as e:
+        print 'zipmerge: target not a zip-file: %s' % target_file
+        raise
 
+    try:        
+        source = zipfile.ZipFile( source_file, 'r' )
+    except Exception as e:
+        print 'zipmerge: source not a zip-file: %s' % source_file
+        raise
+        
     #source supersedes target
     source_files = set( source.namelist() )
     target_files = set( target.namelist() ) - source_files
@@ -225,15 +343,23 @@ def osArch():
         return '32'
 
 def main(mcp_dir):
-    print 'Using mcp dir: %s' % mcp_dir
     print 'Using base dir: %s' % base_dir
-    print("Downloading dependencies...")
-    download_deps( mcp_dir )
+    print 'Using mcp dir: %s (use -m <mcp-dir> to change)' % mcp_dir
+    print 'Preferred architecture: %sbit - preferring %sbit native extraction (use -a 32 or -a 64 to change)' % (preferredarch, preferredarch)
+    if nomerge is True:
+        print 'NO Optifine merging'
+    print("\nDownloading dependencies...")
+    download_deps( mcp_dir, True )
 
-    print("Applying Optifine...")
-    optifine = os.path.join(mcp_dir,"jars","libraries","optifine","OptiFine",of_version,"OptiFine-"+of_version+".jar" )
-    zipmerge( os.path.join( mcp_dir,"jars","versions",mc_version,mc_version+".jar"), optifine )
-
+    if nomerge == False:
+        print("Applying Optifine...")
+        optifine = os.path.join(mcp_dir,"jars","libraries","optifine","OptiFine",of_json_name,"OptiFine-"+of_json_name+".jar" )
+        minecraft_jar = os.path.join( mcp_dir,"jars","versions",mc_version,mc_version+".jar")
+        print ' Merging\n  %s\n into\n  %s' % (optifine, minecraft_jar)
+        zipmerge( minecraft_jar, optifine )
+    else:
+        print("Skipping Optifine merge...")
+    
     print("Decompiling...")
     src_dir = os.path.join(mcp_dir, "src","minecraft")
     if os.path.exists( src_dir ):
@@ -256,9 +382,22 @@ def main(mcp_dir):
 
 if __name__ == '__main__':
     parser = OptionParser()
+    parser.add_option('-o', '--no-optifine', dest='nomerge', default=False, action='store_true', help='If specified, no optifine merge will be carried out')
     parser.add_option('-m', '--mcp-dir', action='store', dest='mcp_dir', help='Path to MCP to use', default=None)
+    parser.add_option('-a', '--architecture', action='store', dest='arch', help='Architecture to use (\'32\' or \'64\'); prefer 32 or 64bit dlls', default=None)
     options, _ = parser.parse_args()
 
+    if not options.arch is None:
+        if options.arch is '32':
+            preferredarch = '32'
+        elif options.arch is '64':
+            preferredarch = '64'
+            
+    if preferredarch is '':
+        preferredarch = osArch()
+        
+    nomerge = options.nomerge
+    
     if not options.mcp_dir is None:
         main(os.path.abspath(options.mcp_dir))
     elif os.path.isfile(os.path.join('..', 'runtime', 'commands.py')):
